renamed files again to a working syntax. updated makefile, ran makefile and everything is running as intended
