+++ bus_procurement_cost/_04_cost_per_bus_cleaner.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from bus_cost_utils import *
+from _bus_cost_utils import GCS_PATH, outlier_flag
 from scipy.stats import zscore
